Treat LONG as meaning a 32 bit type in the C expression evaluator even on 64 bit systems and map incoming variable types to LONG or LONG_LONG as appropriate to their size.

+++ ups/xc_machine.h
@@ -84,8 +84,8 @@ typedef union {
 	char d_char;
 	unsigned short d_ushort;
 	short d_short;
-	unsigned long d_ulong;
-	long d_long;
+	unsigned int d_ulong;
+	int d_long;
 	float d_float;
 	stackword_t d_word;
 #if WANT_LDBL
